+++ src/api/issues/update.rs
@@ -15,6 +15,8 @@ pub struct UpdateIssueBuilder<'octo, 'a, 'b, 'c, 'd, 'e> {
     #[serde(skip_serializing_if = "Option::is_none")]
     state: Option<models::IssueState>,
     #[serde(skip_serializing_if = "Option::is_none")]
+    state_reason: Option<models::issues::IssueStateReason>,
+    #[serde(skip_serializing_if = "Option::is_none")]
     milestone: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     labels: Option<&'e [String]>,
@@ -29,6 +31,7 @@ impl<'octo, 'a, 'b, 'c, 'd, 'e> UpdateIssueBuilder<'octo, 'a, 'b, 'c, 'd, 'e> {
             body: None,
             assignees: None,
             state: None,
+            state_reason: None,
             milestone: None,
             labels: None,
         }
@@ -55,6 +58,15 @@ impl<'octo, 'a, 'b, 'c, 'd, 'e> UpdateIssueBuilder<'octo, 'a, 'b, 'c, 'd, 'e> {
     /// The state of the issue.
     pub fn state(mut self, state: impl Into<models::IssueState>) -> Self {
         self.state = Some(state.into());
+        self
+    }
+
+    /// The state reason of the issue.
+    pub fn state_reason(
+        mut self,
+        state_reason: impl Into<models::issues::IssueStateReason>,
+    ) -> Self {
+        self.state_reason = Some(state_reason.into());
         self
     }
 
@@ -99,6 +111,7 @@ mod tests {
             .title("Updated title")
             .body("New body")
             .state(crate::models::IssueState::Closed)
+            .state_reason(crate::models::issues::IssueStateReason::Completed)
             .milestone(1234u64)
             .assignees(assignees)
             .labels(labels);
@@ -109,6 +122,7 @@ mod tests {
                 "title": "Updated title",
                 "body": "New body",
                 "state": "closed",
+                "state_reason": "completed",
                 "milestone": 1234,
                 "assignees": ["ferris"],
                 "labels": ["help wanted", "good first issue"],